fix: modify file structure
