+++ src/utils.rs
@@ -3,3 +3,4 @@ pub mod decode_jwt;
 pub mod encode_jwt;
 pub mod oauth_authentication;
 pub mod verify_password;
+pub mod verify_recaptcha;
